requirements.txt
aiohttp==1.0.5
astroid==2.0.4
async-timeout==3.0.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
discord.py==0.16.12
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
multidict==4.3.1
pylint==2.1.1
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
websockets==3.4
wrapt==1.10.11